fix: ui for starting and stopping containers
