+++ backend/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2024"
 
 [dependencies]
 axum = "0.8.6"
-bollard = { version = "0.19.3", default-features = false, features = ["pipe"] }
+bollard = { version = "0.20.0", default-features = false, features = ["pipe"] }
 chrono = "0.4.42"
 clap = { version = "4.5.49", features = ["derive", "env"] }
 dotenvy = "0.15.7"
