--- frontend/src/components/StartStop/start_stop.scss
@@ -1,23 +1,0 @@
-@use "../../styles/variables" as *;
-
-.start-stop-container {
-	margin: 0 auto;
-    width: calc(min(1000px, 90%));
-
-	display: flex;
-	flex-direction: column;
-	justify-items: center;
-    
-	& .buttons {
-		display: flex;
-		flex-direction: row;
-		justify-content: center;
-		align-items: center;
-		gap: 1rem;
-		margin-top: 1rem;
-
-		button {
-			font-size: 1.2rem;
-		}
-	}
-}
